refactor: made up my mind on file structure
