+++ src/commands/lyrics.ts
@@ -1,8 +1,8 @@
 import { ChatInputCommandInteraction, EmbedBuilder, SlashCommandBuilder } from 'discord.js';
-import { i18n } from '../src/utils/i18n';
+import { i18n } from '../utils/i18n';
 // @ts-expect-error
 import lyricsFinder from 'lyrics-finder';
-import { bot } from '../index';
+import { bot } from '../../index';
 
 export default {
 	data: new SlashCommandBuilder().setName('lyrics').setDescription(i18n.__('lyrics.description')),
