+++ src/commands/ping.ts
@@ -1,5 +1,5 @@
 import { ChatInputCommandInteraction, SlashCommandBuilder } from 'discord.js';
-import { i18n } from '../src/utils/i18n';
+import { i18n } from '../utils/i18n';
 
 export default {
 	data: new SlashCommandBuilder().setName('ping').setDescription(i18n.__('ping.description')),
