+++ src/sardineCommands/generic/rip.ts
